requirements.txt
python-telegram-bot==20.7
gpt4all==2.7.0
asyncio